Another commit.

+++ src/main/java/io/corkline/model/Range.java
@@ -1,7 +1,7 @@
 package io.corkline.model;
 
-public record Range(double min, double max) {
-    public boolean contains(double value) {
+public record Range(int min, int max) {
+    public boolean contains(int value) {
         return min <= value && value <= max;
     }
 }
